+++ src/components/Chart.vue
@@ -1,5 +1,21 @@
 <template>
-  <iframe :src="link" frameborder="0" ref="frame" class="absolute-top-left absolute-bottom-right" style="width : 100%; height: 100%;"></iframe>
+  <div class="chart">
+    <!--<svg style="width:0; height:0; position:absolute;" aria-hidden="true" focusable="false">
+      <defs>
+        <linearGradient id="btcFill" x1="1" x2="1" y1="0" y2="1">
+          <stop offset="0%" stop-color="#f69119"></stop>
+          <stop offset="100%" stop-color="#ffffff"></stop>
+        </linearGradient>
+      </defs>
+    </svg> -->
+    <trend-chart
+      v-if="dataset.length"
+      :datasets="[{data: dataset, fill: true, className: 'curve-btc'}]"
+      :labels="labels"
+      :min="0"
+      :grid="grid"
+    />
+  </div>
 </template>
 
 <script>
@@ -8,66 +24,23 @@ export default {
   name: 'map-frame',
   data () {
     return {
-      link: 'https://flespi.io/mapview',
-      lastCommand: {},
-      mapInited: false
+      dataset: [],
+      labels: {
+        xLabels: [],
+        yLabels: 5
+      },
+      grid: {
+        verticalLines: true,
+        verticalLinesNumber: 1,
+        horizontalLines: true,
+        horizontalLinesNumber: 1
+      }
     }
   },
   methods: {
-    centerMap (point) {
-      this.lastCommand.centermap = point
-      return this
-    },
-    autobounds (flag) {
-      this.lastCommand.autobounds = flag
-      return this
-    },
-    addConfig (config) {
-      this.lastCommand = merge({}, this.lastCommand, config)
-    },
-    addMarkers (markers) {
-      this.lastCommand.addmarkers = markers
-      return this
-    },
-    /* markers = { 'name': { latlng, direction, title } } */
-    addNamedMarkers (markers) {
-      this.lastCommand.namedmarkers = markers
-      return this
-    },
-    addNamedMarker (markers) {
-      this.lastCommand.namedmarkers = { ...this.lastCommand.namedmarkers, ...markers }
-      return this
-    },
-    addPoints (points) {
-      this.lastCommand.appendpoints = points
-      return this
-    },
-    addRoutes (routes) {
-      this.lastCommand.addgroutes = routes
-      return this
-    },
-    clear (target = 'all') {
-      this.lastCommand.clear = target
-      return this
-    },
-    send () {
-      if (this.$refs.frame && this.mapInited) {
-        this.$refs.frame.contentWindow.postMessage(`MapView|cmd:${JSON.stringify(this.lastCommand)}`, '*')
-        this.lastCommand = {}
-      }
-      return this
-    }
+    setData
   },
   created () {
-    const mapInitHandler = (e) => {
-      if (e.data === 'MapView|state:{"ready": true}') {
-        this.mapInited = true
-        this.$refs.frame.contentWindow.postMessage('MapView|cmd:{"ruler": true, "zoomcontrol": true}', '*')
-        this.send()
-        window.removeEventListener('message', mapInitHandler)
-      }
-    }
-    window.addEventListener('message', mapInitHandler)
   }
 }
 </script>