+++ src/config.js
@@ -941,34 +941,208 @@ export default {
         },
         {
           name: 'event_code',
-          title: 'event',
-          width: 450,
+          title: 'Event',
+          width: 200,
           display: true,
           description: 'Log event code and description'
         },
         {
-          name: 'accepted',
+          name: 'response_code',
+          title: 'Response code',
           width: 100,
           display: true,
-          description: 'Messages count has accepted'
+          description: ''
         },
         {
-          name: 'read',
-          width: 100,
+          name: 'message',
+          width: 200,
           display: true,
-          description: 'Messages count has read'
+          description: 'MQTT message'
         },
         {
-          name: 'rejected',
-          width: 100,
+          name: 'request_body',
+          title: 'Request body',
+          width: 200,
           display: true,
-          description: 'Messages count has rejected'
+          description: 'HTTP request body'
         },
         {
-          name: 'skipped',
+          name: 'response',
           width: 100,
           display: true,
-          description: 'Messages count has skipped'
+          description: 'HTTP server response'
+        }
+      ],
+      actions: [
+        {
+          icon: 'mdi-content-copy',
+          label: 'Copy log object',
+          classes: '',
+          mode: ACTION_MODE_SINGLE,
+          type: 'copy'
+        },
+        {
+          icon: 'mdi-content-copy',
+          label: 'Copy log objects',
+          classes: '',
+          mode: ACTION_MODE_MULTI,
+          type: 'copy'
+        }
+      ],
+      viewConfig: {
+        needShowFilter: true,
+        needShowDateRange: true
+      },
+      theme: {
+        color: 'white',
+        bgColor: 'grey-9',
+        contentInverted: true,
+        controlsInverted: true
+      }
+    }
+  },
+  grants: {
+    label: 'Grants',
+    icon: 'mdi-hand-extended',
+    type: 'viewer',
+    // acl: [{
+    //   name: 'grants',
+    //   methods: ['GET'],
+    //   submodules: [
+    //     { name: 'logs', methods: ['GET'] }
+    //   ]
+    // }],
+    isDrawable: false,
+    logs: {
+      vuexModuleName: 'grantsLogs',
+      emptyState: {
+        label: 'Log entries not found',
+        sublabel: 'If you expect to see the log records here',
+        hints: [
+          { html: "Pick a specific date and time using the <span class='text-bold'>date/time picker</span> on the top right." },
+          { html: "Narrow down the search by specifying the desired parameter values in the <span class='text-bold'>filter control</span>." }
+        ]
+      },
+      cols: [
+        {
+          name: 'timestamp',
+          width: 165,
+          display: true,
+          description: 'Log event time',
+          addition: `${locale.slice(0, 3)}:${locale.slice(3)}`
+        },
+        {
+          name: 'event_code',
+          title: 'Event',
+          width: 200,
+          display: true,
+          description: 'Log event code and description'
+        },
+        {
+          name: 'token_id',
+          title: 'Token id',
+          width: 100,
+          display: true,
+          description: ''
+        },
+        {
+          name: 'origin_id',
+          width: 100,
+          display: true,
+          description: 'ID'
+        },
+        {
+          name: 'host',
+          title: 'Host',
+          width: 200,
+          display: true,
+          description: 'Host'
+        }
+      ],
+      actions: [
+        {
+          icon: 'mdi-content-copy',
+          label: 'Copy log object',
+          classes: '',
+          mode: ACTION_MODE_SINGLE,
+          type: 'copy'
+        },
+        {
+          icon: 'mdi-content-copy',
+          label: 'Copy log objects',
+          classes: '',
+          mode: ACTION_MODE_MULTI,
+          type: 'copy'
+        }
+      ],
+      viewConfig: {
+        needShowFilter: true,
+        needShowDateRange: true
+      },
+      theme: {
+        color: 'white',
+        bgColor: 'grey-9',
+        contentInverted: true,
+        controlsInverted: true
+      }
+    }
+  },
+  realms: {
+    label: 'Realms',
+    icon: 'mdi-folder-account',
+    type: 'viewer',
+    // acl: [{
+    //   name: 'realms',
+    //   methods: ['GET'],
+    //   submodules: [
+    //     { name: 'logs', methods: ['GET'] }
+    //   ]
+    // }],
+    isDrawable: false,
+    logs: {
+      vuexModuleName: 'realmsLogs',
+      emptyState: {
+        label: 'Log entries not found',
+        sublabel: 'If you expect to see the log records here',
+        hints: [
+          { html: "Pick a specific date and time using the <span class='text-bold'>date/time picker</span> on the top right." },
+          { html: "Narrow down the search by specifying the desired parameter values in the <span class='text-bold'>filter control</span>." }
+        ]
+      },
+      cols: [
+        {
+          name: 'timestamp',
+          width: 165,
+          display: true,
+          description: 'Log event time',
+          addition: `${locale.slice(0, 3)}:${locale.slice(3)}`
+        },
+        {
+          name: 'event_code',
+          title: 'Event',
+          width: 200,
+          display: true,
+          description: 'Log event code and description'
+        },
+        {
+          name: 'token_id',
+          title: 'Token id',
+          width: 100,
+          display: true,
+          description: ''
+        },
+        {
+          name: 'origin_id',
+          width: 100,
+          display: true,
+          description: 'ID'
+        },
+        {
+          name: 'host',
+          title: 'Host',
+          width: 200,
+          display: true,
+          description: 'Host'
         }
       ],
       actions: [
@@ -1490,6 +1664,50 @@ export default {
     isDrawable: false,
     messages: {
       vuexModuleName: 'deviceTraffic',
+      emptyState: {
+        label: 'Traffic not found',
+        sublabel: 'If you expect to see the raw messages here',
+        hints: [
+          { html: "Make sure your trackers are correctly pointed to this channel’s <span class='text-bold'>IP:port</span> and they are currently sending messages.", wclass: ['col-12'] },
+          { html: "Pick a specific date and time using the <span class='text-bold'>date/time picker</span> on the top right.", wclass: ['col-12'] },
+          { html: "Narrow down the search by specifying the desired parameter values in the <span class='text-bold'>filter control</span>.", wclass: ['col-12'] }
+        ]
+      },
+      actions: [
+        {
+          icon: 'mdi-content-copy',
+          label: 'copy',
+          classes: '',
+          type: 'copy'
+        }
+      ],
+      viewConfig: {
+        needShowFilter: true
+      },
+      theme: {
+        color: 'white',
+        bgColor: 'grey-9',
+        contentInverted: true,
+        controlsInverted: true
+      }
+    }
+  },
+
+  streamTraffic: {
+    label: 'Stream traffic',
+    path: 'tools/stream-traffic',
+    type: 'tools',
+    icon: 'mdi-download-network-outline',
+    acl: [{
+      name: 'streams',
+      methods: ['GET'],
+      submodules: [
+        { name: 'packets', methods: ['GET'] }
+      ]
+    }],
+    isDrawable: false,
+    messages: {
+      vuexModuleName: 'streamTraffic',
       emptyState: {
         label: 'Traffic not found',
         sublabel: 'If you expect to see the raw messages here',