+++ src/layouts/Index.vue
@@ -83,7 +83,7 @@ export default {
       toolboxMode: undefined,
       connectFlag: false,
       isTabsVisible: true,
-      entityByGroup: ['platform', 'webhooks', 'channels', 'devices', 'groups', 'streams', 'calcs', 'intervals', 'plugins', 'hexViewer', 'trafficViewer', 'deviceTraffic', 'modems', 'containers', 'cdns', 'mqtt', 'mqttClient'],
+      entityByGroup: ['platform', 'webhooks', 'realms', 'grants', 'channels', 'devices', 'groups', 'streams', 'calcs', 'intervals', 'plugins', 'hexViewer', 'trafficViewer', 'deviceTraffic', 'streamTraffic', 'modems', 'containers', 'cdns', 'mqtt', 'mqttClient'],
       isNeedSelect: true,
       entityInited: false,
       isInit: Vue.connector.socket.connected()
@@ -204,6 +204,7 @@ export default {
               result.push('hexViewer')
               result.push('trafficViewer')
               result.push('deviceTraffic')
+              result.push('streamTraffic')
               break
             }
             case 'storage': {
@@ -219,6 +220,8 @@ export default {
             case 'platform': {
               result.push('platform')
               result.push('webhooks')
+              result.push('grants')
+              result.push('realms')
               break
             }
           }