+++ src/pages/viewer/Channels.vue
@@ -578,10 +578,10 @@ export default {
         this.$emit('update:settings', { type: 'ENTITY_VIEW_SETTINGS_CHANGE', opt: { entity: this.entityName }, value: { ratio: r } })
         this.$nextTick(() => {
           if (+this.size[0] && this.active) {
-            this.$refs.logs.resetParams()
+            this.$refs.logs && this.$refs.logs.resetParams()
           }
           if (+this.size[1] && this.active) {
-            this.$refs.messages.resetParams()
+            this.$refs.messages && this.$refs.messages.resetParams()
           }
         })
       }